feat: support for gpt
