+++ src/options/generate/templates_user.py
@@ -90,8 +90,6 @@ You must provide the complete file with the exact same syntax to wrap the code.'
 template_generate_executor = PromptTemplate.from_template(
     general_guidelines_string + '''
 
-{code_files_wrapped}
-
 Write the executor called '{microservice_name}'. The name is very important to keep.
 It matches the following description: '{microservice_description}'.
 It will be tested with the following scenario: '{test_description}'.
@@ -103,9 +101,32 @@ Have in mind that d.uri is never a path to a local file. It is always a url.
 
 Your approach:
 1. Identify the core challenge when implementing the executor.
-2. Think about solutions for these challenges including the usage of gpt via "from gpt_3_5_turbo_api import GPT_3_5_Turbo_API"
+2. Think about solutions for these challenges including the usage of gpt via "gpt_3_5_turbo_api"
 3. Decide for one of the solutions.
 4. Write the code for the executor. Don't write code for the test.
+If you decided to use gpt, then the executor must include the following code:
+import os
+import openai
+openai.api_key = os.getenv("OPENAI_API_KEY")
+
+class GPT_3_5_Turbo_API:
+    def __init__(self, system: str = ''):
+        self.system = system
+
+    def __call__(self, prompt: str) -> str:
+        response = openai.ChatCompletion.create(
+            model="gpt-3.5-turbo",
+            messages=[{{
+                "role": 'system',
+                "content": self.system
+            }}, {{
+                "role": 'user',
+                "content": prompt
+            }}]
+        )
+        return response.choices[0]['message']['content']
+
+
 ''' + '\n' + template_code_wrapping_string
 )
 
@@ -158,8 +179,7 @@ It is important to make sure that all libs are installed that are required by th
 Usually libraries are installed with apt-get.
 Be aware that the machine the docker container is running on does not have a GPU - only CPU.
 Add the config.yml file to the Dockerfile.
-Add the gpt_3_5_turbo_api.py file to the Dockerfile.
-Note that the Dockerfile only has access to the files: microservice.py, requirements.txt, config.yml, test_microservice.py and gpt_3_5_turbo_api.py.
+Note that the Dockerfile only has access to the files: microservice.py, requirements.txt, config.yml and test_microservice.py.
 The base image of the Dockerfile is FROM jinaai/jina:3.14.1-py39-standard.
 The entrypoint is ENTRYPOINT ["jina", "executor", "--uses", "config.yml"].
 Make sure the all files are in the /workdir.
@@ -275,7 +295,7 @@ template_generate_playground = PromptTemplate.from_template(
 Create a playground for the executor {microservice_name} using streamlit.
 The playground must look like it was made by a professional designer.
 All the ui elements are well thought out to make them visually appealing and easy to use.
-The playground contains emojis that fit the theme of the playground.
+The playground contains many emojis that fit the theme of the playground and has an emoji as favicon.
 This is an example how you can connect to the executor assuming the document (d) is already defined:
 ```
 from jina import Client, Document, DocumentArray
